Better separation of deferred graph construction and the overloaded blaze Kernel
