+++ blaze/deferred.py
@@ -1,12 +1,23 @@
 # -*- coding: utf-8 -*-
 
 """
-Deferred array node.
+Deferred array node and kernel application.
 """
 
 from __future__ import print_function, division, absolute_import
-from blaze.expr import dump
 
+from itertools import chain
+
+import blaze
+from blaze.datashape import coretypes as T
+from blaze.expr import dump
+from blaze.expr.context import merge
+
+from blaze.py2help import dict_iteritems
+
+#------------------------------------------------------------------------
+# Deferred Array
+#------------------------------------------------------------------------
 
 class Deferred(object):
     """
@@ -41,3 +52,59 @@ class Deferred(object):
             pass
 
         return dump(term, ipython=ipython)
+
+
+#------------------------------------------------------------------------
+# Application
+#------------------------------------------------------------------------
+
+def apply_kernel(kernel, *args, **kwargs):
+    """
+    Apply blaze kernel `kernel` to the given arguments.
+
+    Returns: a Deferred node representation the delayed computation
+    """
+    from .expr import construct
+
+    # -------------------------------------------------
+    # Merge input contexts
+
+    args, kwargs = blaze_args(args, kwargs)
+    ctxs = collect_contexts(chain(args, kwargs.values()))
+    ctx = merge(ctxs)
+
+    # -------------------------------------------------
+    # Find match to overloaded function
+
+    match, args = kernel.dispatcher.lookup_dispatcher(args, kwargs,
+                                                    ctx.constraints)
+
+    # -------------------------------------------------
+    # Construct graph
+
+    term = construct.construct(kernel, ctx, match.func, match.dst_sig, args)
+    return blaze.Deferred(term.dshape, (term, ctx))
+
+#------------------------------------------------------------------------
+# Utils
+#------------------------------------------------------------------------
+
+def blaze_args(args, kwargs):
+    """Build blaze arrays from inputs to a blaze kernel"""
+    args = [make_blaze(a) for a in args]
+    kwargs = dict((v, make_blaze(k)) for k, v in dict_iteritems(kwargs))
+    return args, kwargs
+
+def make_blaze(value):
+    if not isinstance(value, (blaze.Deferred, blaze.Array)):
+        dshape = T.typeof(value)
+        if not dshape.shape:
+            value = [value]
+        value = blaze.Array([value], dshape)
+    return value
+
+def collect_contexts(args):
+    for term in args:
+        if term.expr:
+            t, ctx = term.expr
+            yield ctx
